Move suites tests to their own file to make panic triage easier..
The triage process excludes files named "_test.rs".
